refactor: extract message selection state into useMessageSelection composable
